requirements.txt
Django==4.2.5
factory-boy==3.3.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
django-widget-tweaks==1.5.0
openpyxl==3.1.2